Moved main to user folder
