+++ src/App.jsx
@@ -17,7 +17,7 @@ import ForgotPassword from './components/pages/login/ForgotPassword'
 import VerifyCode from './components/pages/login/VerifyCode'
 import AlternativesPage from './components/pages/AlternativesPage'
 import SearchResults from './components/pages/SearchResults'
-import Main from './components/pages/main/Main'
+import Main from './components/pages/user/main/Main'
 import Prescription from './components/pages/Prescription'
 import SignUp from './components/pages/signUp/SignUp'
 import InsuranceCompanies from './components/pages/insurance/InsuranceCompanies'
