+++ src/components/pages/user/main/AddPrescription.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 /*       Icons       */
-import PrescriptionImage from '../../../assets/mainPage/PrescriptionImage.svg'
+import PrescriptionImage from '../../../../assets/mainPage/PrescriptionImage.svg'
 
 const AddPrescription = () => {
   return (
