+++ src/components/pages/user/main/ImageSlider.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect, useRef } from 'react'
-import RightArrow from '../../../assets/mainPage/RightArrow.svg'
-import LeftArrow from '../../../assets/mainPage/LeftArrow.svg'
-import Image_1 from '../../../assets/mainPage/Image_1.png'
-import Image_2 from '../../../assets/mainPage/Image_2.png'
-import Image_3 from '../../../assets/mainPage/Image_3.png'
-import Image_4 from '../../../assets/mainPage/Image_4.png'
-import Image_5 from '../../../assets/mainPage/Image_5.png'
+import RightArrow from '../../../../assets/mainPage/RightArrow.svg'
+import LeftArrow from '../../../../assets/mainPage/LeftArrow.svg'
+import Image_1 from '../../../../assets/mainPage/Image_1.png'
+import Image_2 from '../../../../assets/mainPage/Image_2.png'
+import Image_3 from '../../../../assets/mainPage/Image_3.png'
+import Image_4 from '../../../../assets/mainPage/Image_4.png'
+import Image_5 from '../../../../assets/mainPage/Image_5.png'
 import './Dot.css'
 import { FaCircle } from 'react-icons/fa'
 /*       css       */
